# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "FOODIE-FUND",
    "description": "眾資成城-在你心愛的餐廳成為合夥人"
  },
  "servers": [
    {
      "url": "https://foodiefund.onrender.com",
      "description": "Render"
    },
    {
      "url": "http://127.0.0.1:2330",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/v1/api/admin/account/": {
      "get": {
        "tags": [
          "會員管理"
        ],
        "summary": "會員清單",
        "description": "會員清單",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "description": "管理身分, user,admin",
            "enum": [
              "user",
              "admin",
              "undefined"
            ],
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isBlackListed",
            "in": "query",
            "description": "1:搜尋黑名單 0:不是黑名單, 預設空直為搜尋全部",
            "enum": [
              1,
              0,
              "undefined"
            ],
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "keyWord",
            "in": "query",
            "description": "關鍵字fuzzy[email,phone,address,dateOfBirth], 預設空直為搜尋全部",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeSort",
            "in": "query",
            "description": "排序遠到近,進到遠default = desc",
            "enum": [
              "asc",
              "desc"
            ],
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "清單顯示比數,default=10",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "顯示第幾頁資料default=1",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/api/admin/account/blackList/{id}": {
      "patch": {
        "tags": [
          "會員管理"
        ],
        "summary": "設定黑名單",
        "description": "設定黑名單",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "使用者id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": ""
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "66d0c762273627e056be5238"
                        },
                        "name": {
                          "type": "string",
                          "example": "Lobinda"
                        },
                        "email": {
                          "type": "string",
                          "example": "lobinda123@test.com"
                        },
                        "phone": {
                          "type": "string",
                          "example": "0987654321"
                        },
                        "address": {
                          "type": "string",
                          "example": "地球某個角落"
                        },
                        "date_of_birth": {
                          "type": "string",
                          "example": "2006-08-18T00:00:00.000Z"
                        },
                        "role": {
                          "type": "string",
                          "example": "user"
                        },
                        "remarks": {
                          "type": "string",
                          "example": ""
                        },
                        "isBlackListed": {
                          "type": "boolean",
                          "example": false
                        },
                        "id": {
                          "type": "string",
                          "example": "66d0c762273627e056be5238"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": ""
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "66d0c762273627e056be5238"
                        },
                        "name": {
                          "type": "string",
                          "example": "Lobinda"
                        },
                        "email": {
                          "type": "string",
                          "example": "lobinda123@test.com"
                        },
                        "phone": {
                          "type": "string",
                          "example": "0987654321"
                        },
                        "address": {
                          "type": "string",
                          "example": "地球某個角落"
                        },
                        "date_of_birth": {
                          "type": "string",
                          "example": "2006-08-18T00:00:00.000Z"
                        },
                        "role": {
                          "type": "string",
                          "example": "user"
                        },
                        "remarks": {
                          "type": "string",
                          "example": ""
                        },
                        "isBlackListed": {
                          "type": "boolean",
                          "example": false
                        },
                        "id": {
                          "type": "string",
                          "example": "66d0c762273627e056be5238"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Mail not verified"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Mail not verified"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "會員資料",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isBlackListed": {
                    "type": "boolean",
                    "enum": [
                      "1",
                      "0"
                    ],
                    "example": "1"
                  }
                },
                "required": [
                  "isBlackListed"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/api/admin/account/roleCharacter/{id}": {
      "patch": {
        "tags": [
          "會員管理"
        ],
        "summary": "設定權限",
        "description": "設定權限",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "使用者id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": ""
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "66d0c762273627e056be5238"
                        },
                        "name": {
                          "type": "string",
                          "example": "Lobinda"
                        },
                        "email": {
                          "type": "string",
                          "example": "lobinda123@test.com"
                        },
                        "phone": {
                          "type": "string",
                          "example": "0987654321"
                        },
                        "address": {
                          "type": "string",
                          "example": "地球某個角落"
                        },
                        "date_of_birth": {
                          "type": "string",
                          "example": "2006-08-18T00:00:00.000Z"
                        },
                        "role": {
                          "type": "string",
                          "example": "user"
                        },
                        "remarks": {
                          "type": "string",
                          "example": ""
                        },
                        "isBlackListed": {
                          "type": "boolean",
                          "example": false
                        },
                        "id": {
                          "type": "string",
                          "example": "66d0c762273627e056be5238"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": ""
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "66d0c762273627e056be5238"
                        },
                        "name": {
                          "type": "string",
                          "example": "Lobinda"
                        },
                        "email": {
                          "type": "string",
                          "example": "lobinda123@test.com"
                        },
                        "phone": {
                          "type": "string",
                          "example": "0987654321"
                        },
                        "address": {
                          "type": "string",
                          "example": "地球某個角落"
                        },
                        "date_of_birth": {
                          "type": "string",
                          "example": "2006-08-18T00:00:00.000Z"
                        },
                        "role": {
                          "type": "string",
                          "example": "user"
                        },
                        "remarks": {
                          "type": "string",
                          "example": ""
                        },
                        "isBlackListed": {
                          "type": "boolean",
                          "example": false
                        },
                        "id": {
                          "type": "string",
                          "example": "66d0c762273627e056be5238"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Mail not verified"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Mail not verified"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "會員資料",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "example": "user,admin"
                  }
                },
                "required": [
                  "role"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}